requirements.txt
openai==0.24.0
numpy==1.21.6
python-dotenv==0.21.0
transformers==4.26.1